Added files via upload
It is desirable to have granular reporting of add-to-cart and purchase activity, but it is undesirable to make multiple tracking calls for multi-item events.  In this case, a “bundled” call format can be used. In this format parameter names are appended with and index number enclosed in square braces (i.e. sku[0]). If the bundled format is used the indexing should start at zero.  All calls support bundling, though it is typically only useful for purchases.